+++ src/main/java/org/tinyradius/client/handler/PromiseAdapter.java
@@ -28,7 +28,7 @@ public class PromiseAdapter extends MessageToMessageCodec<RadiusResponse, Pendin
 
     private static final Logger logger = LogManager.getLogger();
 
-    private static final int PROXY_STATE = 33;
+    private static final byte PROXY_STATE = 33;
 
     private final AtomicInteger proxyIndex = new AtomicInteger(1);
