+++ src/test/java/org/tinyradius/attribute/Ipv6PrefixAttributeTest.java
@@ -14,37 +14,37 @@ class Ipv6PrefixAttributeTest {
     @Test
     void minAttributeLength() {
         // min
-        final Ipv6PrefixAttribute prefixAttribute = new Ipv6PrefixAttribute(dictionary, -1, 97, new byte[2]);
+        final Ipv6PrefixAttribute prefixAttribute = new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, new byte[2]);
         assertEquals(2, prefixAttribute.getValue().length);
 
         // min-1
         IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new Ipv6PrefixAttribute(dictionary, -1, 97, new byte[1]));
+                () -> new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, new byte[1]));
         assertTrue(exception.getMessage().toLowerCase().contains("should be 2-18 octets"));
     }
 
     @Test
     void maxAttributeLength() {
         // max
-        final Ipv6PrefixAttribute prefixAttribute = new Ipv6PrefixAttribute(dictionary, -1, 97, new byte[18]);
+        final Ipv6PrefixAttribute prefixAttribute = new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, new byte[18]);
         assertEquals(2, prefixAttribute.getValue().length); // prefix-length set to 0, so everything else is trimmed
 
         // max+1
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new Ipv6PrefixAttribute(dictionary, -1, 97, new byte[20]));
+                () -> new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, new byte[20]));
         assertTrue(exception.getMessage().toLowerCase().contains("should be 2-18 octets"));
     }
 
     @Test
     void stringOk() {
-        final Ipv6PrefixAttribute attribute = new Ipv6PrefixAttribute(dictionary, -1, 97, "2001:db8:ac10:fe01:0:0:0:0/64");
+        final Ipv6PrefixAttribute attribute = new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, "2001:db8:ac10:fe01:0:0:0:0/64");
         assertEquals("2001:db8:ac10:fe01:0:0:0:0/64", attribute.getValueString());
     }
 
     @Test
     void stringEmpty() {
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new Ipv6PrefixAttribute(dictionary, -1, 97, ""));
+                () -> new Ipv6PrefixAttribute(dictionary, -1, (byte) 97, ""));
         assertTrue(exception.getMessage().toLowerCase().contains("invalid ipv6 prefix"));
     }
 
@@ -52,7 +52,7 @@ class Ipv6PrefixAttributeTest {
     void bitsOutsidePrefixLengthIsZero() {
         // string constructor
         final IllegalArgumentException e1 = assertThrows(IllegalArgumentException.class,
-                () -> Attributes.create(dictionary, -1, 97, "fe80:fe80:fe80:fe80:fe80:fe80:fe80:fe80/64"));
+                () -> Attributes.create(dictionary, -1, (byte) 97, "fe80:fe80:fe80:fe80:fe80:fe80:fe80:fe80/64"));
 
         assertTrue(e1.getMessage().toLowerCase().contains("bits outside of the prefix-length must be zero"));
 
@@ -63,7 +63,7 @@ class Ipv6PrefixAttributeTest {
         bytes[3] = (byte) 0xff; // violates rfc requirement that "Bits outside of the Prefix-Length, if included, must be zero."
 
         final IllegalArgumentException e2 = assertThrows(IllegalArgumentException.class,
-                () -> create(dictionary, -1, 97, bytes));
+                () -> create(dictionary, -1, (byte) 97, bytes));
 
         assertTrue(e2.getMessage().toLowerCase().contains("bits outside of the prefix-length must be zero"));
     }
@@ -76,7 +76,7 @@ class Ipv6PrefixAttributeTest {
         bytes[3] = 0; // set 8 bits to 0, but also part of prefix
         bytes[4] = 0; // empty octet as padding
 
-        final RadiusAttribute attribute = create(dictionary, -1, 97, bytes);
+        final RadiusAttribute attribute = create(dictionary, -1, (byte) 97, bytes);
         assertEquals("fe00:0:0:0:0:0:0:0/16", attribute.getValueString());
     }
 
@@ -86,12 +86,12 @@ class Ipv6PrefixAttributeTest {
         final byte[] bytes = new byte[4]; // prefix capacity 2
         bytes[1] = 16; // 16 bits require 2 bytes
 
-        final RadiusAttribute attribute = create(dictionary, -1, 97, bytes);
+        final RadiusAttribute attribute = create(dictionary, -1, (byte) 97, bytes);
         assertEquals("0:0:0:0:0:0:0:0/16", attribute.getValueString());
 
         bytes[1] = 17; // 17 bits require 3 bytes;
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> create(dictionary, -1, 97, bytes));
+                () -> create(dictionary, -1, (byte) 97, bytes));
         assertTrue(exception.getMessage().toLowerCase().contains("actual byte array only has space for 16 bits"));
         assertTrue(exception.getMessage().toLowerCase().contains("prefix-length declared 17 bits"));
     }
